api/contact/serializers.py
from rest_framework import serializers


class ContactMessageSerializer(serializers.Serializer):
    name = serializers.CharField(max_length=200)
    email = serializers.EmailField()
    content = serializers.CharField(max_length=3000)
